refactor page builder into hooks and components
