+++ packages/ui/src/components/cms/page-builder/components/PageCanvas.tsx
@@ -2,10 +2,10 @@ import { SortableContext, rectSortingStrategy } from "@dnd-kit/sortable";
 import type { CSSProperties, DragEvent } from "react";
 import { Fragment } from "react";
 import type { PageComponent } from "@acme/types";
-import CanvasItem from "./CanvasItem";
+import CanvasItem from "../CanvasItem";
 import type { Locale } from "@/i18n/locales";
-import type { Action } from "./state";
-import { cn } from "../../../utils/style";
+import type { Action } from "../state";
+import { cn } from "../../../../utils/style";
 
 interface Props {
   components: PageComponent[];
