+++ packages/ui/src/components/cms/page-builder/index.ts
@@ -34,12 +34,13 @@ export { default as CustomHtmlEditor } from "./CustomHtmlEditor";
 export { default as useMediaLibrary } from "./useMediaLibrary";
 export { useArrayEditor } from "./useArrayEditor";
 export { default as CanvasItem } from "./CanvasItem";
-export { default as Palette } from "./Palette";
+export { default as Palette } from "./components/Palette";
 export { default as PageBuilder } from "./PageBuilder";
-export { default as PageToolbar } from "./PageToolbar";
-export { default as PageCanvas } from "./PageCanvas";
-export { default as PageSidebar } from "./PageSidebar";
+export { default as PageToolbar } from "./components/PageToolbar";
+export { default as PageCanvas } from "./components/PageCanvas";
+export { default as PageSidebar } from "./components/PageSidebar";
 export { historyStateSchema, reducer } from "./state";
-export { default as usePageBuilderDrag } from "./usePageBuilderDrag";
+export { default as usePageBuilderState } from "./hooks/usePageBuilderState";
+export { default as usePageBuilderDnD } from "./hooks/usePageBuilderDnD";
 export { default as useCanvasDrag } from "./useCanvasDrag";
 export { default as useCanvasResize } from "./useCanvasResize";
